requirements.txt
Django==5.0.5
djangorestframework